devops/spotnet/margin/margin_app/pyproject.toml
[tool.poetry]
package-mode = false
name = "starkhero"
version = "0.1.0"
description = ""
authors = ["djeck1432 <oleh.kostashchuk@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
loguru = "^0.7.3"
pydantic-settings = "^2.7.1"
asyncpg = "^0.30.0"
pydantic = {extras = ["email"], version = "^2.10.5"}
greenlet = "^3.1.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
argon2-cffi = "^23.1.0"
alembic-postgresql-enum = "^1.7.0"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
pytest-mock = "^3.14.0"
pytest-env = "^1.1.5"
pytest-cov = "^6.0.0"
trio = "^0.29.0"
pyjwt = "^2.10.1"
faker = "^37.0.2"
sendgrid = "^6.9.0"
requests = "^2.32.3"
starlette = "^0.46.1"
itsdangerous = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
